fix: 正确构建

+++ scripts/build.ts
@@ -8,11 +8,12 @@ const argv = minimist(process.argv.slice(2))
 const target = argv._
 
 async function resolveBuild() {
+  await buildWeb('../packages/cli', true)
   const opts = await buildOptions()
   for (let i = 0; i < target.length; i++) {
     const [input, output] = opts[target[i]] ? opts[target[i]]() : opts._normal(target[i])
     const distPath = output.dir
-    if (fs.existsSync(distPath)) {
+    if (target[i] === 'core' && fs.existsSync(distPath)) {
       const files = fs.readdirSync(distPath)
       for (let i = 0; i < files.length; i++)
         fs.unlinkSync(`${distPath}/${files[i]}`)
@@ -20,7 +21,6 @@ async function resolveBuild() {
     const bundle = await rollup(input)
     await bundle.write(output)
   }
-  await buildWeb('../packages/cli', true)
 }
 
 resolveBuild()
